Moved numpy utils out of the way
