+++ src/bot.py
@@ -11,8 +11,7 @@ from util.sequence import Sequence, ControlStep
 from util.spikes import SpikeWatcher
 from util.vec import Vec3
 
-import numpy as np
-from util.np_utils import a3v, closest_to, linear_predict
+# Would you to use numpy utilities? Check out the np_util folder!
 
 class MyBot(BaseAgent):
 
@@ -61,26 +60,6 @@ class MyBot(BaseAgent):
             elif self.spike_watcher.carry_duration > 3:
                 return SimpleControllerState(use_item=True)
 
-        # Example of numpy utilities:
-        # Get all car positions into an array.
-        car_positions = []
-        for car in packet.game_cars[:packet.num_cars]:
-            car_positions.append(a3v(car.physics.location))
-        car_positions_arr = np.vstack(car_positions)
-        # Get the ball position.
-        ball_pos = a3v(packet.game_ball.physics.location)
-        # Find the closest car to the ball.
-        closest_index = closest_to(ball_pos, car_positions_arr)
-        # Prepare render message.
-        closest_text = f'The bot with index {closest_index} is closest to the ball.'
-
-        # Render the linear prediction of the closest car.
-        closest_vel = a3v(packet.game_cars[closest_index].physics.velocity)
-        prediction = linear_predict(car_positions[closest_index], closest_vel, packet.game_info.seconds_elapsed, 2.0)
-        self.renderer.begin_rendering('prediction')
-        self.renderer.draw_polyline_3d(prediction['pos'][::30], self.renderer.pink())
-        self.renderer.end_rendering()
-
         # The rest of this code just ball chases.
         # Find the direction of our car using the Orientation class
         car_orientation = Orientation(my_car.physics.rotation)
@@ -93,7 +72,7 @@ class MyBot(BaseAgent):
         self.controller_state.throttle = 1.0
         self.controller_state.steer = -1 if steer_correction_radians > 0 else 1.0
 
-        draw_debug(self.renderer, [goal_text, closest_text])
+        draw_debug(self.renderer, [goal_text])
 
         return self.controller_state
